商品管理ページにおいて、商品削除実行時に、imgディレクトリに保存の同商品画像も併せて削除するよう変更

+++ bungushop/new_account.php
@@ -29,7 +29,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $user_name = get_post_data('user_name');
     $passwd = get_post_data('passwd');
     $mail = get_post_data('mail');
-    $sex = get_post_data('sex');
+    $sex = (int)get_post_data('sex');
     $birthdate = get_post_data('birth');
 
     //
